Split workload set tests into two classes
