+++ test/dotnet-MsiInstallation.Tests/WorkloadSetTests.cs
@@ -9,29 +9,10 @@ using Microsoft.NET.Sdk.WorkloadManifestReader;
 
 namespace Microsoft.DotNet.MsiInstallerTests
 {
-    public class WorkloadSetTests : VMTestBase
+    public class WorkloadSetTests : WorkloadSetTestsBase
     {
-        readonly string SdkTestingDirectory = @"C:\SdkTesting";
-
-
-        Lazy<Dictionary<string, string>> _testWorkloadSetVersions;
-        string WorkloadSetVersion1 => _testWorkloadSetVersions.Value["version1"];
-        string WorkloadSetVersion2 => _testWorkloadSetVersions.Value["version2"];
-        string WorkloadSetPreviousBandVersion => _testWorkloadSetVersions.Value.GetValueOrDefault("previousbandversion", "8.0.204");
-
         public WorkloadSetTests(ITestOutputHelper log) : base(log)
         {
-            _testWorkloadSetVersions = new Lazy<Dictionary<string, string>>(() =>
-            {
-                string remoteFilePath = @"c:\SdkTesting\workloadsets\testworkloadsetversions.json";
-                var versionsFile = VM.GetRemoteFile(remoteFilePath);
-                if (!versionsFile.Exists)
-                {
-                    throw new FileNotFoundException($"Could not find file {remoteFilePath} on VM");
-                }
-
-                return JsonSerializer.Deserialize<Dictionary<string, string>>(versionsFile.ReadAllText());
-            });
         }
 
         [Fact]
@@ -39,7 +20,7 @@ namespace Microsoft.DotNet.MsiInstallerTests
         {
             InstallSdk();
 
-            VM.CreateRunCommand("dotnet", "workload", "update", "--include-previews")
+            CreateInstallingCommand("dotnet", "workload", "update")
                 .Execute()
                 .Should()
                 .PassWithoutWarning();
@@ -48,7 +29,7 @@ namespace Microsoft.DotNet.MsiInstallerTests
 
             AddNuGetSource(@"c:\SdkTesting\WorkloadSets");
 
-            VM.CreateRunCommand("dotnet", "workload", "update", "--include-previews")
+            CreateInstallingCommand("dotnet", "workload", "update")
                 .Execute()
                 .Should()
                 .PassWithoutWarning();
@@ -57,42 +38,6 @@ namespace Microsoft.DotNet.MsiInstallerTests
 
             newRollback.ManifestVersions.Should().BeEquivalentTo(originalRollback.ManifestVersions);
 
-        }
-
-        void UpdateAndSwitchToWorkloadSetMode(out string updatedWorkloadVersion, out WorkloadSet rollbackAfterUpdate)
-        {
-            var featureBand = new SdkFeatureBand(SdkInstallerVersion).ToStringWithoutPrerelease();
-            var originalWorkloadVersion = GetWorkloadVersion();
-            originalWorkloadVersion.Should().StartWith($"{featureBand}-manifests.");
-
-            VM.CreateRunCommand("dotnet", "workload", "update", "--include-previews")
-                .Execute()
-                .Should()
-                .PassWithoutWarning();
-
-            rollbackAfterUpdate = GetRollback();
-            updatedWorkloadVersion = GetWorkloadVersion();
-            updatedWorkloadVersion.Should().StartWith($"{featureBand}-manifests.");
-            updatedWorkloadVersion.Should().NotBe(originalWorkloadVersion);
-
-            GetUpdateMode().Should().Be("manifests");
-
-            VM.CreateRunCommand("dotnet", "workload", "config", "--update-mode", "workload-set")
-                .WithDescription("Switch mode to workload-set")
-                .Execute()
-                .Should()
-                .PassWithoutWarning();
-
-            GetWorkloadVersion().Should().Be(updatedWorkloadVersion);
-
-            var expectedMessage = "Workloads are configured to install and update using workload versions, but none were found. Run \"dotnet workload restore\" to install a workload version.";
-
-            GetDotnetInfo().Should().Contain(expectedMessage)
-                .And.NotContain("(not installed)");
-            GetDotnetWorkloadInfo().Should().Contain(expectedMessage)
-                .And.NotContain("(not installed)");
-
-            GetUpdateMode().Should().Be("workload-set");
         }
 
         [Fact]
@@ -104,7 +49,7 @@ namespace Microsoft.DotNet.MsiInstallerTests
 
             AddNuGetSource(@"c:\SdkTesting\WorkloadSets");
 
-            VM.CreateRunCommand("dotnet", "workload", "update", "--include-previews")
+            CreateInstallingCommand("dotnet", "workload", "update")
                 .Execute().Should().PassWithoutWarning();
 
             GetWorkloadVersion().Should().Be(WorkloadSetVersion2);
@@ -121,7 +66,7 @@ namespace Microsoft.DotNet.MsiInstallerTests
             UpdateAndSwitchToWorkloadSetMode(out string updatedWorkloadVersion, out WorkloadSet rollbackAfterUpdate);
 
             //  Use a nonexistant source because there may be a valid workload set available on NuGet.org
-            VM.CreateRunCommand("dotnet", "workload", "update", "--include-previews", "--source", @"c:\SdkTesting\EmptySource")
+            CreateInstallingCommand("dotnet", "workload", "update", "--source", @"c:\SdkTesting\EmptySource")
                 .Execute()
                 .Should()
                 .PassWithoutWarning();
@@ -154,7 +99,7 @@ namespace Microsoft.DotNet.MsiInstallerTests
 
             AddNuGetSource(@"c:\SdkTesting\WorkloadSets");
 
-            VM.CreateRunCommand("dotnet", "workload", "update", "--version", versionToInstall, "--include-previews")
+            CreateInstallingCommand("dotnet", "workload", "update", "--version", versionToInstall)
                 .Execute()
                 .Should()
                 .PassWithoutWarning();
@@ -172,7 +117,7 @@ namespace Microsoft.DotNet.MsiInstallerTests
 
             GetWorkloadVersion().Should().Be(versionToInstall);
 
-            VM.CreateRunCommand("dotnet", "workload", "update", "--include-previews")
+            CreateInstallingCommand("dotnet", "workload", "update")
                 .Execute()
                 .Should()
                 .PassWithoutWarning();
@@ -189,7 +134,7 @@ namespace Microsoft.DotNet.MsiInstallerTests
 
             var workloadVersionBeforeUpdate = GetWorkloadVersion();
 
-            VM.CreateRunCommand("dotnet", "workload", "update", "--version", unavailableWorkloadSetVersion, "--include-previews")
+            CreateInstallingCommand("dotnet", "workload", "update", "--version", unavailableWorkloadSetVersion)
                 .Execute()
                 .Should()
                 .Fail()
@@ -215,7 +160,7 @@ namespace Microsoft.DotNet.MsiInstallerTests
 
             var workloadVersionBeforeUpdate = GetWorkloadVersion();
 
-            VM.CreateRunCommand("dotnet", "workload", "update", "--source", @"c:\SdkTesting\workloadsets", "--include-previews")
+            CreateInstallingCommand("dotnet", "workload", "update", "--source", @"c:\SdkTesting\workloadsets")
                 .Execute()
                 .Should()
                 .Fail();
@@ -236,7 +181,7 @@ namespace Microsoft.DotNet.MsiInstallerTests
 
             var workloadVersionBeforeUpdate = GetWorkloadVersion();
 
-            VM.CreateRunCommand("dotnet", "workload", "update", "--version", WorkloadSetVersion2, "--source", @"c:\SdkTesting\workloadsets", "--include-previews")
+            CreateInstallingCommand("dotnet", "workload", "update", "--version", WorkloadSetVersion2, "--source", @"c:\SdkTesting\workloadsets")
                 .Execute()
                 .Should()
                 .Fail();
@@ -248,150 +193,6 @@ namespace Microsoft.DotNet.MsiInstallerTests
                 .PassWithoutWarning();
 
             GetWorkloadVersion().Should().Be(workloadVersionBeforeUpdate);
-        }
-
-        void SetupWorkloadSetInGlobalJson(out WorkloadSet originalRollback)
-        {
-            InstallSdk();
-
-            var versionToUpdateTo = WorkloadSetVersion2;
-
-            string originalVersion = GetWorkloadVersion();
-
-            originalRollback = GetRollback(SdkTestingDirectory);
-
-            VM.WriteFile("C:\\SdkTesting\\global.json", @$"{{""sdk"":{{""workloadVersion"":""{versionToUpdateTo}""}}}}").Execute().Should().PassWithoutWarning();
-
-            GetWorkloadVersion(SdkTestingDirectory).Should().Be(versionToUpdateTo + " (not installed)");
-            GetDotnetInfo(SdkTestingDirectory).Should().Contain("Workload version:  " + versionToUpdateTo + " (not installed)")
-                .And.Contain($@"Workload version {versionToUpdateTo}, which was specified in C:\SdkTesting\global.json, was not found");
-            GetDotnetWorkloadInfo(SdkTestingDirectory).Should().Contain("Workload version: " + versionToUpdateTo + " (not installed)")
-                .And.Contain($@"Workload version {versionToUpdateTo}, which was specified in C:\SdkTesting\global.json, was not found");
-
-            // The version should have changed but not yet the manifests. Since we expect both, getting the rollback should fail.
-            var result = VM.CreateRunCommand("dotnet", "workload", "update", "--print-rollback")
-               .WithWorkingDirectory(SdkTestingDirectory)
-               .WithIsReadOnly(true)
-               .Execute();
-
-            result.Should().Fail();
-            result.StdErr.Should().Contain("FileNotFoundException");
-            result.StdErr.Should().Contain(versionToUpdateTo);
-
-            AddNuGetSource(@"C:\SdkTesting\workloadsets", SdkTestingDirectory);
-        }
-
-        [Fact]
-        public void RestoreWorkloadSetViaGlobalJson()
-        {
-            InstallSdk();
-
-            var testProjectFolder = Path.Combine(SdkTestingDirectory, "ConsoleApp");
-            VM.CreateRunCommand("dotnet", "new", "console", "-o", "ConsoleApp")
-                .WithWorkingDirectory(SdkTestingDirectory)
-                .Execute().Should().PassWithoutWarning();
-
-            SetupWorkloadSetInGlobalJson(out var originalRollback);
-
-            VM.CreateRunCommand("dotnet", "workload", "restore")
-                .WithWorkingDirectory(testProjectFolder)
-                .Execute().Should().PassWithoutWarning();
-
-            GetWorkloadVersion(SdkTestingDirectory).Should().Be(WorkloadSetVersion2);
-
-            GetRollback(SdkTestingDirectory).Should().NotBe(originalRollback);
-        }
-
-        [Theory]
-        [InlineData("update")]
-        [InlineData("install")]
-        public void UseGlobalJsonToSpecifyWorkloadSet(string command)
-        {
-            SetupWorkloadSetInGlobalJson(out var originalRollback);
-
-            string[] args = command.Equals("install") ? ["dotnet", "workload", "install", "aspire"] : ["dotnet", "workload", command];
-            VM.CreateRunCommand(args).WithWorkingDirectory(SdkTestingDirectory).Execute().Should().PassWithoutWarning();
-            GetRollback(SdkTestingDirectory).Should().NotBe(originalRollback);
-        }
-
-        [Fact]
-        public void DotnetInfoWithGlobalJson()
-        {
-            InstallSdk();
-
-            //  Install a workload before setting up global.json.  Commands like "dotnet workload --info" were previously crashing if global.json specified a workload set that wasn't installed
-            InstallWorkload("aspire", skipManifestUpdate: true);
-
-            SetupWorkloadSetInGlobalJson(out _);
-        }
-
-        [Fact]
-        public void InstallWithGlobalJsonAndSkipManifestUpdate()
-        {
-            SetupWorkloadSetInGlobalJson(out var originalRollback);
-
-            VM.CreateRunCommand("dotnet", "workload", "install", "aspire", "--skip-manifest-update")
-                .WithWorkingDirectory(SdkTestingDirectory)
-                .Execute().Should().Fail()
-                .And.HaveStdErrContaining("--skip-manifest-update")
-                .And.HaveStdErrContaining(Path.Combine(SdkTestingDirectory, "global.json"));
-        }
-
-        [Fact]
-        public void InstallWithVersionAndSkipManifestUpdate()
-        {
-            InstallSdk();
-
-            VM.CreateRunCommand("dotnet", "workload", "install", "aspire", "--skip-manifest-update", "--version", WorkloadSetVersion1)
-                .Execute().Should().Fail()
-                .And.HaveStdErrContaining("--skip-manifest-update")
-                .And.HaveStdErrContaining("--sdk-version");
-        }
-
-        [Fact]
-        public void InstallWithVersionWhenPinned()
-        {
-            InstallSdk();
-
-            AddNuGetSource(@"c:\SdkTesting\WorkloadSets");
-
-            string originalVersion = GetWorkloadVersion();
-            originalVersion.Should().NotBe(WorkloadSetVersion1);
-
-            VM.CreateRunCommand("dotnet", "workload", "update", "--version", WorkloadSetVersion1, "--include-previews")
-                .Execute().Should().PassWithoutWarning();
-
-            GetWorkloadVersion().Should().Be(WorkloadSetVersion1);
-
-            VM.CreateRunCommand("dotnet", "workload", "install", "aspire", "--version", WorkloadSetVersion2)
-                .Execute().Should().PassWithoutWarning();
-
-            GetWorkloadVersion().Should().Be(WorkloadSetVersion2);
-        }
-
-        [Fact]
-        public void InstallWithGlobalJsonWhenPinned()
-        {
-            SetupWorkloadSetInGlobalJson(out var originalRollback);
-
-            //AddNuGetSource(@"c:\SdkTesting\WorkloadSets");
-
-            string originalVersion = GetWorkloadVersion();
-            originalVersion.Should().NotBe(WorkloadSetVersion1);
-
-            VM.CreateRunCommand("dotnet", "workload", "update", "--version", WorkloadSetVersion1, "--include-previews")
-                .Execute().Should().PassWithoutWarning();
-
-            GetWorkloadVersion().Should().Be(WorkloadSetVersion1);
-
-            VM.CreateRunCommand("dotnet", "workload", "install", "aspire")
-                .WithWorkingDirectory(SdkTestingDirectory)
-                .Execute().Should().PassWithoutWarning();
-
-            GetWorkloadVersion(SdkTestingDirectory).Should().Be(WorkloadSetVersion2);
-
-            GetRollback(SdkTestingDirectory).Should().NotBe(originalRollback);
-
         }
 
         [Fact]
@@ -410,7 +211,7 @@ namespace Microsoft.DotNet.MsiInstallerTests
                     VM.CreateRunCommand("cmd", "/c", "ren", @$"c:\SdkTesting\WorkloadSets\Microsoft.NET.Workloads.{sdkFeatureBand}.*.{packageVersion}.nupkg", $"Microsoft.NET.Workloads.{sdkFeatureBand}.*.{packageVersion}.bak"))
                 .Execute().Should().PassWithoutWarning();
 
-            VM.CreateRunCommand("dotnet", "workload", "update", "--include-previews")
+            CreateInstallingCommand("dotnet", "workload", "update")
                 .Execute().Should().PassWithoutWarning();
 
             GetWorkloadVersion().Should().Be(WorkloadSetVersion1);
@@ -452,7 +253,7 @@ namespace Microsoft.DotNet.MsiInstallerTests
             AddNuGetSource(@"c:\SdkTesting\WorkloadSets");
 
             //  Update to latest workload set version
-            VM.CreateRunCommand("dotnet", "workload", "update", "--include-previews")
+            CreateInstallingCommand("dotnet", "workload", "update")
                 .Execute().Should().PassWithoutWarning();
 
             GetWorkloadVersion().Should().Be(WorkloadSetVersion2);
@@ -475,12 +276,12 @@ namespace Microsoft.DotNet.MsiInstallerTests
             VM.WriteFile("C:\\SdkTesting\\global.json", @$"{{""sdk"":{{""workloadVersion"":""{WorkloadSetVersion1}""}}}}").Execute().Should().PassWithoutWarning();
 
             //  Install pinned version
-            VM.CreateRunCommand("dotnet", "workload", "update", "--include-previews")
+            CreateInstallingCommand("dotnet", "workload", "update")
                .WithWorkingDirectory(SdkTestingDirectory)
                .Execute().Should().PassWithoutWarning();
 
             //  Update globally installed version to later version
-            VM.CreateRunCommand("dotnet", "workload", "update", "--include-previews")
+            CreateInstallingCommand("dotnet", "workload", "update")
                .Execute().Should().PassWithoutWarning();
 
             //  Check workload versions in global context and global.json directory
@@ -495,7 +296,7 @@ namespace Microsoft.DotNet.MsiInstallerTests
             VM.WriteFile("C:\\SdkTesting\\global.json", "{}").Execute().Should().PassWithoutWarning();
 
             //  Run workload update to do a GC
-            VM.CreateRunCommand("dotnet", "workload", "update", "--include-previews")
+            CreateInstallingCommand("dotnet", "workload", "update")
                .Execute().Should().PassWithoutWarning();
 
             //  Workload set 1 should have been GC'd
@@ -557,7 +358,7 @@ namespace Microsoft.DotNet.MsiInstallerTests
             var searchResultWorkloadSet = WorkloadSet.FromDictionaryForJson(JsonSerializer.Deserialize<Dictionary<string, string>>(searchResultJson["manifestVersions"]), new SdkFeatureBand(SdkInstallerVersion));
 
             //  Update to the workload set version we got the search info from so we can check to see if the manifest versions match what we expect
-            VM.CreateRunCommand("dotnet", "workload", "update", "--include-previews", "--version", WorkloadSetVersion2)
+            CreateInstallingCommand("dotnet", "workload", "update", "--version", WorkloadSetVersion2)
                 .Execute()
                 .Should()
                 .PassWithoutWarning();
@@ -565,61 +366,5 @@ namespace Microsoft.DotNet.MsiInstallerTests
             GetRollback().ManifestVersions.Should().BeEquivalentTo(searchResultWorkloadSet.ManifestVersions);
         }
 
-        string GetWorkloadVersion(string workingDirectory = null)
-        {
-            var result = VM.CreateRunCommand("dotnet", "workload", "--version")
-                .WithWorkingDirectory(workingDirectory)
-                .WithIsReadOnly(true)
-                .Execute();
-
-            result.Should().PassWithoutWarning();
-
-            return result.StdOut;
-        }
-
-        string GetDotnetInfo(string workingDirectory = null)
-        {
-            var result = VM.CreateRunCommand("dotnet", "--info")
-                .WithWorkingDirectory(workingDirectory)
-                .WithIsReadOnly(true)
-                .Execute();
-
-            result.Should().PassWithoutWarning();
-
-            return result.StdOut;
-        }
-
-        string GetDotnetWorkloadInfo(string workingDirectory = null)
-        {
-            var result = VM.CreateRunCommand("dotnet", "workload", "--info")
-                .WithWorkingDirectory(workingDirectory)
-                .WithIsReadOnly(true)
-                .Execute();
-
-            result.Should().PassWithoutWarning();
-
-            return result.StdOut;
-        }
-
-        string GetUpdateMode()
-        {
-            var result = VM.CreateRunCommand("dotnet", "workload", "config", "--update-mode")
-                .WithIsReadOnly(true)
-                .Execute();
-
-            result.Should().PassWithoutWarning();
-
-            return result.StdOut;
-        }
-
-        void AddNuGetSource(string source, string directory = null)
-        {
-            VM.CreateRunCommand("dotnet", "nuget", "add", "source", source)
-                .WithWorkingDirectory(directory)
-                .WithDescription($"Add {source} to NuGet.config")
-                .Execute()
-                .Should()
-                .PassWithoutWarning();
-        }
     }
 }
